requirements.txt
# Core AI and ML dependencies
ollama==0.1.7
langchain==0.1.0
langchain-community==0.0.10
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
pandas==2.1.3

# Database connectors
pymongo==4.6.0
psycopg2-binary==2.9.9
pymysql==1.1.0
sqlalchemy==2.0.23
redis==5.0.1

# Monitoring and metrics
prometheus-client==0.19.0
psutil==5.9.6
py-cpuinfo==9.0.0

# Web interface and API
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
gradio==4.0.2
websockets>=10.0,<12.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.2
schedule==1.2.0
click==8.1.7
rich==13.7.0

# Database monitoring specific
sqlparse==0.4.4
sqlalchemy-utils==0.41.1
alembic==1.13.0

# Configuration and validation
pyyaml==6.0.1
jsonschema==4.20.0 